added signal handling

+++ nipype/pipeline/plugins/multiproc.py
@@ -7,35 +7,46 @@ Support for child processes running as non-daemons based on
 http://stackoverflow.com/a/8963618/1183453
 """
 
-from multiprocessing import (Process, Pool, cpu_count, pool)
+from multiprocessing import (Process, Pool, cpu_count, pool, Value)
 from copy import deepcopy
 from traceback import format_exception
 import sys
 import signal
+import numpy as np
 from .base import (DistributedPluginBase, report_crash, report_nodes_not_run)
 from ..engine import MapNode
 from ...utils.misc import str2bool
-import numpy as np
 from ... import logging
-logger = logging.getLogger('workflow')
+
+LOGGER = logging.getLogger('workflow')
+INTERRUPT = None
 
 
 def run_node(args):
+    """ Picklable function to execute a node """
+    global INTERRUPT
     jobid, node, updatehash = args
-    logger.info('[Starting] Job %d %s' % (jobid, str(node._id)))
     jres = {'result': None}
+
+    if INTERRUPT.value:
+        LOGGER.error('[Skipped] Job %d %s' % (jobid, str(node._id)))
+        jres['traceback'] = ['Cancelled upon user request']
+        return jres
+
+    LOGGER.info('[Starting] Job %d %s' % (jobid, str(node.fullname)))
     try:
         jres['result'] = node.run(updatehash=updatehash)
-        logger.info('[Terminated] Job %d %s' % (jobid, str(node._id)))
-    except Exception:
-        logger.error('[Failed] Job %d %s' % (jobid, str(node._id)))
-        etype, ev, etr = sys.exc_info()
-        jres['traceback'] = format_exception(etype, ev, etr)
-        jres['result'] = node.result
+        LOGGER.info('[Terminated] Job %d %s' % (jobid, str(node._id)))
     except KeyboardInterrupt:
-        logger.error('Processing interrupted by user.')
-        etype, ev, etr = sys.exc_info()
-        jres['traceback'] = format_exception(etype, ev, etr)
+        LOGGER.error('[Cancelled] Job %d %s' % (jobid, str(node._id)))
+        with INTERRUPT.value.get_lock():
+            INTERRUPT.value = True
+        etype, eval, etr = sys.exc_info()
+        jres['traceback'] = format_exception(etype, eval, etr)
+    except Exception:
+        LOGGER.error('[Failed] Job %d %s' % (jobid, str(node._id)))
+        etype, eval, etr = sys.exc_info()
+        jres['traceback'] = format_exception(etype, eval, etr)
         jres['result'] = node.result
 
     return (jobid, jres)
@@ -76,13 +87,21 @@ class MultiProcPlugin(DistributedPluginBase):
 
     """
 
-    def _init_worker(self):
+    def kill_all(self, *args):
+        self.pool.close()
+        self.pool.terminate()
+        self.pool.join()
+        raise KeyboardInterrupt()
+
+    def _init_worker(self, args):
         """
         Ignore interrupt signal, handle in main thread
         https://noswap.com/blog/python-multiprocessing-keyboardinterrupt
         """
-        pass
         # signal.signal(signal.SIGINT, signal.SIG_IGN)
+        global INTERRUPT
+        INTERRUPT = args
+        LOGGER.info('Started worker, INTERRUPT=%s' % INTERRUPT.value)
 
     def _start_pool(self):
         """
@@ -94,7 +113,7 @@ class MultiProcPlugin(DistributedPluginBase):
             # run the execution using the appropriate pool subclass
             self.pool = NonDaemonPool(**self._poolcfg) \
                 if self._non_daemon else Pool(**self.poolcfg)
-            logger.info('Started new %s pool with %d processes' %
+            LOGGER.info('Started new %s pool with %d processes' %
                         ('non-daemon' if self._non_daemon else 'daemon',
                          self._poolcfg['processes']))
         except TypeError:
@@ -112,6 +131,8 @@ calling-helper-functions-when-using-apply-asyncs-callback
         self._results = {}   # Save results here
         self._notrun = []    # Gather errored nodes here
 
+        interrupt = Value('b', False)
+
         # Initialize settings, using dic.get we define defaults
         if plugin_args is None:
             plugin_args = {}
@@ -119,12 +140,14 @@ calling-helper-functions-when-using-apply-asyncs-callback
         self._poolcfg = dict(
             processes=plugin_args.get('n_proc', cpu_count()),
             initializer=self._init_worker,
+            initargs=(interrupt, ),
             maxtasksperchild=plugin_args.get('maxtasksperchild', 50))
         self._non_daemon = plugin_args.get('non_daemon', True)
 
     def _submit_jobs(self, jobids, graph=None, updatehash=False):
+        global INTERRUPT
         jobids = np.atleast_1d(jobids).tolist()
-        logger.info('Submitting %s' % str(jobids))
+        LOGGER.info('Submitting %s' % str(jobids))
         jobargs = []
         active = []
         for jobid in jobids:
@@ -138,9 +161,13 @@ calling-helper-functions-when-using-apply-asyncs-callback
                 active.append(jobid)
                 jobargs.append((jobid, node, updatehash,))
 
-        logger.info('Current pool is %s' % str(active))
+        LOGGER.info('Current pool is %s' % str(active))
         self._start_pool()
         cur_batch = self.pool.map(run_node, jobargs)
+
+        if INTERRUPT.value:
+            raise KeyboardInterrupt()
+
         self.pool.close()
         self.pool.terminate()
         del self.pool
@@ -185,7 +212,7 @@ calling-helper-functions-when-using-apply-asyncs-callback
 
         if num_jobs > 0:
             # send all available jobs
-            logger.info('Check %d jobs and submit' % len(jobids[:slots]))
+            LOGGER.info('Check %d jobs and submit' % len(jobids[:slots]))
             for jobid in jobids[:slots]:
                 if isinstance(self.procs[jobid], MapNode):
                     try:
@@ -206,7 +233,7 @@ calling-helper-functions-when-using-apply-asyncs-callback
                 continue_with_submission = True
                 if str2bool(self.procs[jobid].config['execution']
                             ['local_hash_check']):
-                    logger.info('checking hash locally')
+                    LOGGER.info('checking hash locally')
                     hash_exists, _, _, _ = self.procs[jobid].hash_exists()
                     overwrite = getattr(
                         self.procs[jobid], 'overwrite', False)
@@ -219,7 +246,7 @@ calling-helper-functions-when-using-apply-asyncs-callback
                         self.proc_done[jobid] = True
                         self._task_finished_cb(jobid)
                         self._remove_node_dirs()
-                        logger.info(('Node %s (%d) is cached or does not require being run') %
+                        LOGGER.info(('Node %s (%d) is cached or does not require being run') %
                                     (self.procs[jobid], jobid))
 
                 if continue_with_submission:
@@ -241,7 +268,7 @@ calling-helper-functions-when-using-apply-asyncs-callback
         """
         Executes a pre-defined pipeline using distributed approaches
         """
-        logger.info('Running in parallel.')
+        LOGGER.info('Running in parallel.')
         self._config = config
         # Generate appropriate structures for worker-manager model
         self._generate_dependency_list(graph)
@@ -256,7 +283,7 @@ calling-helper-functions-when-using-apply-asyncs-callback
             undone = np.logical_not(self.proc_done).astype(np.uint8)
             jobdeps = (self.depidx.sum(axis=0) == 0).__array__()
             jobids = np.flatnonzero(undone * jobdeps)
-            logger.info(('Polling [%02d]: Jobs ready/remaining/total='
+            LOGGER.info(('Polling [%02d]: Jobs ready/remaining/total='
                          '%d/%d/%d') % (it, len(jobids), undone.sum(),
                                         len(self.proc_done)))
 
@@ -264,7 +291,7 @@ calling-helper-functions-when-using-apply-asyncs-callback
                 jobids = jobids[:self.max_jobs]
 
             if len(jobids) == 0:
-                logger.info('No more ready jobs, exit')
+                LOGGER.info('No more ready jobs, exit')
                 break
 
             self._send_procs_to_workers(jobids, updatehash=updatehash,
